Re-introducing WITH-TRANSACTION (change previously reverted because of package problems, which should now be fixed).
Original commit message:

I thought Weblocks should export a WITH-TRANSACTION macro to go with the
other functions in this file.  Implementing it, I noticed a bug in
DYNAMIC-TRANSACTION: it would commit, rather than rollback, on a nonlocal
exit out of the body if that exit wasn't done by signalling an error
(i.e. was done by THROW, RETURN, or GO).  I fixed this.  I guessed
that this was the "unique non-local exit unwind behavior" being warned
about, and removed the warning.